Better approach to resolving relative file paths
Replaced strange approach to resolving relative folder  / file paths
with an approach that can consistently get the application root path
using the AppDomain. This avoids problems when the HttpContext is not
available.

Better approach to resolving relative file paths

Application was using a strange workaround to avoid problems when the
HttpContext was not available for resolving relative file paths.
Replaced this with a more consistent approach using the current app
domain.