Added master page to search.aspx
Also removed unsafe code that was displaying SQL on the page. While this
SQL was hidden for non-admin users, it was still available to anyone who
looked at the HTML for the page. Also, this SQL was being saved back to
the server when the user chose to save the search.  It would have been
very easy for someone to inject whatever SQL they wanted through this
save search feature.